projects/py_projects/pvs/requirements-test.txt
mypy==0.700
pylint==2.3.1
pytest==4.3.1
pytest-aiohttp==0.3.0
sqlalchemy-stubs==0.1
